LayoutTests/inspector/dom/highlightNodeList-expected.txt

Tests for the DOM.highlightNodeList command.


PASS: Page should have a subframe.

== Running test suite: DOM.highlightNodeList
-- Running test case: DOM.highlightNodeList.CheckEmptyHighlight
PASS: Highlight should not exist.

-- Running test case: DOM.highlightNodeList.emptyNodeList
PASS: Should highlight 0 element(s).

-- Running test case: DOM.highlightNodeList.mainFrameNode
PASS: Should highlight 3 element(s).
Highlighted Elements:
{
  "tagName": "div",
  "idValue": "",
  "size": {
    "width": 500,
    "height": 500
  },
  "role": ""
}
{
  "tagName": "div",
  "idValue": "",
  "classes": [
    "class-one"
  ],
  "size": {
    "width": 10,
    "height": 20
  },
  "role": ""
}
{
  "tagName": "div",
  "idValue": "id-one",
  "classes": [
    "class-two"
  ],
  "size": {
    "width": 100,
    "height": 200
  },
  "role": ""
}

-- Running test case: DOM.highlightNodeList.childFrameNode
PASS: Should highlight 2 element(s).
Highlighted Elements:
{
  "tagName": "div",
  "idValue": "id-one",
  "size": {
    "width": 150,
    "height": 250
  },
  "role": ""
}
{
  "tagName": "div",
  "idValue": "",
  "size": {
    "width": 269,
    "height": 0
  },
  "role": ""
}

-- Running test case: DOM.highlightNodeList.invalidNodeId
PASS: Should highlight 0 element(s).

-- Running test case: DOM.highlightNodeList.nonIntegerNodeId
PASS: Should produce an error.
Error: Invalid nodeIds item type. Expecting integer types.
